feat: parallelize best build process
